Route Metadata (Route/Middleware Enhancement)
This commit/PR adds the ability to add arbitrary metadata to an HTTP route which is then passed down the call-chain to middleware and the handler function.

This allows middleware to recieve route-specific information and can be used in several ways such as:
- "Roles" defined on a route that influences middleware processing
- Using metadata to conditionally skip the middleware on specific HTTP endpoints

To do this the following was done:
- Added a helper function (`_create_route_entry`) to `RestAPIEventHandler` for getting route_entry
- Added the ability to pass "metadata" kwargs
- `RestAPIEventHandler`'s call function now also looks at RouteEntry and
  also injects the metadata object into the request.

This allows more complex middleware to be created that needs
route-specific data, such as "roles" or even allows the ability to use
route metadata to conditionally run middlewares on routes.